AIchatter/requirement.txt
AudioSegment
SpeechRecognition
torch
gTTS
click
numpy
configparser
openai-whisper
PyAudio